twitter kadai 1

+++ app/Http/Controllers/TasksController.php
@@ -15,14 +15,35 @@ class TasksController extends Controller
      */
     public function index()
     {
+        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->get();
+
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
+        else{
+        // Welcomeビューでそれらを表示
+       return view('welcome');
+        }
+    }
+        
+        
+        
         // メッセージ一覧を取得
-        $tasks = Task::all();
+      //  $tasks = Task::all();
 
         // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
-    }
+        //return view('tasks.index', [
+          //  'tasks' => $tasks,
+        //]);
+    
 
     /**
      * Show the form for creating a new resource.
@@ -54,10 +75,21 @@ class TasksController extends Controller
         ]);
         
         // メッセージを作成
-        $task = new task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+       // $task = new task;
+        //$task->status = $request->status;    // 追加
+        //$task->content = $request->content;
+        //$task->save();
+
+
+
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,'status' => $request->status
+        ]);
+
+
+
+
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -74,10 +106,24 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+             // メッセージを削除
+             return view('tasks.show', [
+            'task' => $task,
+        ]);
+        }    
+        else{
+            return redirect('/');
+        
+        }
+        
+
+/*
         return view('tasks.show', [
             'task' => $task,
         ]);
+  */      
     }
 
     /**
@@ -119,8 +165,21 @@ class TasksController extends Controller
         $task->content = $request->content;
         $task->save();
 
+       
+
+
         // トップページへリダイレクトさせる
         return redirect('/');
+        
+        
+        
+        
+        
+        
+        
+        
+        
+        
     }
 
     /**
@@ -131,11 +190,18 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
+        
+
+        
         // idの値でメッセージを検索して取得
         $task = task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
-
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+             // メッセージを削除
+             $task->delete();
+        }
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
